requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
requests==2.26.0
python-dotenv==0.19.0
Wand